Rounding text coordinates to int to avoid blurring

+++ SnowyPeak.Duality.Plugins.YAUI/Controls/Button.cs
@@ -57,8 +57,8 @@ namespace SnowyPeak.Duality.Plugins.YAUI.Controls
 				}
 
 				canvas.DrawText(_fText,
-					textPosition.X,
-					textPosition.Y,
+                    (int)textPosition.X,
+                    (int)textPosition.Y,
 					zOffset + (INNER_ZOFFSET * 2),
 					null,
 					this.TextConfiguration.Alignment);
